Renamed Exclude to ExcludeFeature to avoid a conflict with the Exclude procedure which excludes an item from a set.
Added ObjectStates to support IsInspecting

Added TgBase.Assign to allow object assignement.

Added TgBase.Inspect to set the inspecting state, so an object manually creating an object property in a getter will know if the problem is really being accessed, or just queried.

Added TgRecordProperty record  to cache the Getter, Setter and Validator methods of a record property. Also added G.RecordProperty to return the record given an RTTIProperty.