Test/TestgCore.pas:
Added ValidatePhone validation attribute and attached in to the TPhoneString type.

Changed Ancestor of TBase, TBase2 and TBase3 to TgObject.

Added Required attribute to DateProperty, DateTimeProperty, IntegerProperty, ObjectProperty,  StringProperty, and String5 to test the Required attribute.

Added a ValidateRequired test.

gCore.Pas:

Moved PopulateDefaultValues, AutoCreate and Destroy from TgBase to TgObject.

Changed how property validation works. Instead of storing the RTTIProperty object in the attribute, the PropertyValidationAttributes cache stores a record containing both the attribute and the RTTIProperty. This allows us to include a validation on a type as well as a property. The property indirectly gets the validation as a result of it being of a type that has validation.

Added the Required validation attribute.

Fixed the G.InitializeCompositeProperties method. It will probably need more work, though.