Added TgSerializer.CreateAndDeserialize to allow for the deserialization of an object whose class is not known until runtime.
Added G.ClassByName to convert a qualified class name string into a class.

Modified TgBase.FriendlyName to replace '<' and '>' characters with a '_' character. This had to be done since these symbols are used in generic type names and we use the friendly name for filenames and xml tag names.

Modified G.InitializePersistenceManagers to use CreateAndDeserialize to allow for runtime creation of a persistent manager class whose name is contained in a file.