I added a nested list class type to TgPersistenceManagerFile to store the LastID for a TgIDObject. This allows the persistence manager to assign the next ID if it's value is zero when saved.
Added a DeserializeUnpublishedProperty method to the XML and JSON Serialization Helpers to allow TgList's unpublished list property to deserialize. The previous implementation did not allow for a list to contain additional serializable properties (as was needed for the LastID from the list mentioned in the preceding paragraph).

Added HasIdentity and RemoveIdentity to TgIdentity object to test for a value, and remove the value in the ID property, respectively.