Added PathName an OwnerProperty to TgBase.
PathName returns the property names, delimited by a dot, from the highest owner to self.

OwnerProperty returns the TRTTIProperty of the object property in the owner referring to self.

Added TgIdentityList.CurrentKey. This will be used as an index into the list. Unlike CurrentIndex, when CurrentKey gets set to an ID of an IdendityObject, the persistence manager returns back to the list the single object matching the ID.