Added TgIdentityList.ExtendedWhere to filter the list based on ownership context.
Modified TgBase.Assign to allow assignment of an object property when the source is not owned, but the destination is. Also made sure the property is writable before assigning if neither the source or destination is not owned.

Fixed a bug in TgList.Filter. Deletion of the last item causes an automatic Previous, so I removed the Previous call when a Delete takes place.

Added tests.