Added G.IdentityListProperties
Removed the saved current index from an identity list's enumerator.

Modified Required.Execute to ensure the object exists, and if it's an identity object, that its reference exists.

Modified TgIdentityObject.GetCanSave to return a positive result if there is no identity.

Modified TgIdentityObject.SetID to set IsLoaded to False, and deactivate identity lists if the object had a previous identity.