Fixed a bug in TgSerializationHelperXMLBase.Deserialize where an empty value caused an access violation.
Added TgPersistenceManagerFile.AssignChanged to only assign values that changed. This gets called by the TgPersistenceManagerFile.SaveObject method.

Fixed a bug in TgIdentityList.ExtendedWhere.

Fixed a bug in Required.Execute that tested if an identity object exists by loading in place. This caused the object to initialize its OriginalValues. This interfered with the modified test, and caused the ID not to get written out. I changed the implementation to test for the object's existence by creating a new object (instead of using the live one) and loading it.

In Test/TestgCore.pas, I added a SaveChanges test to prove only changed values get saved.