Added TgHTMLString type
Added TgBase.Properties array to return the RTTIProperty for a property path.

Added TgHTMLExpressionEvaluator and a companion EvalHTML function for use with the HTML template evaluator. It's the same as the Eval function, but has an OUT parameter to let you know if there's HTML in the result. If HTML exists in the result, the HTML template evaluator has to create a new DOM, pass in the HTML and evaluate it against the destination DOM.